[#2565] sync with upstream
git-svn-id: svn://localhost/svn/uclinux-dist/trunk@5333 3e2a5524-39e8-4cb9-84a9-c49da69cba3a

+++ user/netflash/crypto.h
@@ -19,7 +19,7 @@
  * host to target system which causes a few problems!
  */
 struct header {
-	unsigned long magic;
+	unsigned int magic;
 	unsigned char md5[MD5_DIGEST_LENGTH];
 	unsigned char aeskey[AESKEYSIZE];
 	unsigned char flags;
