+++ user/squid/helpers/external_acl/winbind_group/wb_check_group.c
@@ -364,7 +364,7 @@ main (int argc, char *argv[])
 	if ((p = strchr(buf, '\r')) != NULL)
 	    *p = '\0';		/* strip \r */
 
-	debug("Got '%s' from Squid (length: %d).\n",buf,strlen(buf));
+	debug("Got '%s' from Squid (length: %d).\n",buf,(int)strlen(buf));
 	
 	if (buf[0] == '\0') {
 	    warn("Invalid Request\n");
