+++ user/squid/src/fs/diskd/store_diskd.h
@@ -59,7 +59,7 @@ typedef struct _diomsg {
     int seq_no;
     void *callback_data;
     int size;
-    int offset;
+    off_t offset;
     int status;
     int shm_offset;
 } diomsg;
@@ -95,8 +95,8 @@ extern char *storeDiskdDirFullPath(SwapDir * SD, sfileno filn, char *fullpath);
 extern void storeDiskdDirUnlinkFile(SwapDir *, sfileno);
 extern void storeDiskdDirReplAdd(SwapDir *, StoreEntry *);
 extern void storeDiskdDirReplRemove(StoreEntry *);
-extern void storeDiskdShmPut(SwapDir *, off_t);
-extern void *storeDiskdShmGet(SwapDir *, off_t *);
+extern void storeDiskdShmPut(SwapDir *, int);
+extern void *storeDiskdShmGet(SwapDir *, int *);
 extern void storeDiskdHandle(diomsg * M);
 extern int storeDiskdDirCallback(SwapDir *);
