+++ user/squid/src/repl/heap/store_heap_replacement.c
@@ -1,6 +1,6 @@
 
 /*
- * $Id$
+ * $Id: store_heap_replacement.c,v 1.7.2.1 2005/02/13 21:23:19 serassio Exp $
  *
  * DEBUG: section 20    Storage Manager Heap-based replacement
  * AUTHOR: John Dilley
@@ -79,7 +79,7 @@ HeapKeyGen_StoreEntry_LFUDA(void *entry, double age)
 	tie = 1.0 - exp((double) (e->lastref - squid_curtime) / 86400.0);
     key = age + (double) e->refcount - tie;
     debug(81, 3) ("HeapKeyGen_StoreEntry_LFUDA: %s refcnt=%d lastref=%ld age=%f tie=%f -> %f\n",
-	storeKeyText(e->hash.key), (int)e->refcount, e->lastref, age, tie, key);
+	storeKeyText(e->hash.key), (int) e->refcount, (long int) e->lastref, age, tie, key);
     if (e->mem_obj && e->mem_obj->url)
 	debug(81, 3) ("HeapKeyGen_StoreEntry_LFUDA: url=%s\n",
 	    e->mem_obj->url);
@@ -115,7 +115,7 @@ HeapKeyGen_StoreEntry_GDSF(void *entry, double age)
     double tie = (e->lastref > 1) ? (1.0 / e->lastref) : 1.0;
     key = age + ((double) e->refcount / size) - tie;
     debug(81, 3) ("HeapKeyGen_StoreEntry_GDSF: %s size=%f refcnt=%d lastref=%ld age=%f tie=%f -> %f\n",
-	storeKeyText(e->hash.key), size, (int)e->refcount, e->lastref, age, tie, key);
+	storeKeyText(e->hash.key), size, (int) e->refcount, (long int) e->lastref, age, tie, key);
     if (e->mem_obj && e->mem_obj->url)
 	debug(81, 3) ("HeapKeyGen_StoreEntry_GDSF: url=%s\n",
 	    e->mem_obj->url);
